requirements.txt
# ==================================================
# 核心套件（PDF 處理）
# ==================================================
PyMuPDF==1.26.4              # PDF 讀取和處理
camelot-py==1.0.9            # 表格擷取（gemini.py 需要）

# ==================================================
# 圖像處理
# ==================================================
opencv-python==4.12.0.88     # 圖像處理和輪廓偵測
Pillow==11.0.0               # 圖片讀寫

# ==================================================
# 資料處理
# ==================================================
numpy==2.1.2                 # 數值運算
pandas==2.3.3                # 資料框架（處理表格）

# ==================================================
# Camelot 相依套件
# ==================================================
pdfminer.six==20250506       # PDF 文字擷取
pypdfium2==4.30.0            # PDF 渲染
chardet==5.2.0               # 字元編碼偵測
click==8.3.0                 # 命令列工具

# ==================================================
# 其他工具
# ==================================================
python-dateutil==2.9.0.post0 # 日期處理
pytz==2025.2                 # 時區處理
psycopg2-binary==2.9.11
pydantic==2.11.7
google-generativeai==0.8.5   #Google AI相關
sentence-transformers==5.1.2 #Google AI相關
pdfplumber==0.11.7
torch==2.9.0                 #透過顯卡運行bge-m3